+++ internal/impl/pure/output_broker.go
@@ -83,9 +83,14 @@ The greedy pattern results in higher output throughput at the cost of potentiall
 				Advanced().
 				Default(1),
 			service.NewStringEnumField(boFieldPattern,
-				"fan_out", "fan_out_fail_fast", "fan_out_sequential", "fan_out_sequential_fail_fast", "round_robin", "greedy").
+				string(service.OutputBrokerPatternFanOut),
+				string(service.OutputBrokerPatternFanOutFailFast),
+				string(service.OutputBrokerPatternFanOutSequential),
+				string(service.OutputBrokerPatternFanOutSequentialFailFast),
+				string(service.OutputBrokerPatternRoundRobin),
+				string(service.OutputBrokerPatternGreedy)).
 				Description("The brokering pattern to use.").
-				Default("fan_out"),
+				Default(string(service.OutputBrokerPatternFanOut)),
 			service.NewOutputListField(boFieldOutputs).
 				Description("A list of child outputs to broker."),
 			service.NewBatchPolicyField(boFieldBatching),
@@ -142,10 +147,10 @@ func brokerOutputFromParsed(conf *service.ParsedConfig, res *service.Resources) 
 		}
 	}
 
-	_, isRetryWrapped := map[string]struct{}{
-		"fan_out":            {},
-		"fan_out_sequential": {},
-	}[pattern]
+	_, isRetryWrapped := map[service.OutputBrokerPatternType]struct{}{
+		service.OutputBrokerPatternFanOut:           {},
+		service.OutputBrokerPatternFanOutSequential: {},
+	}[service.OutputBrokerPatternType(pattern)]
 
 	var outputs []output.Streamed
 	{
@@ -193,14 +198,14 @@ func brokerOutputFromParsed(conf *service.ParsedConfig, res *service.Resources) 
 	}
 
 	var b output.Streamed
-	switch pattern {
-	case "fan_out", "fan_out_fail_fast":
+	switch service.OutputBrokerPatternType(pattern) {
+	case service.OutputBrokerPatternFanOut, service.OutputBrokerPatternFanOutFailFast:
 		b, err = newFanOutOutputBroker(outputs)
-	case "fan_out_sequential", "fan_out_sequential_fail_fast":
+	case service.OutputBrokerPatternFanOutSequential, service.OutputBrokerPatternFanOutSequentialFailFast:
 		b, err = newFanOutSequentialOutputBroker(outputs)
-	case "round_robin":
+	case service.OutputBrokerPatternRoundRobin:
 		b, err = newRoundRobinOutputBroker(outputs)
-	case "greedy":
+	case service.OutputBrokerPatternGreedy:
 		b, err = newGreedyOutputBroker(outputs)
 	default:
 		return nil, fmt.Errorf("broker pattern was not recognised: %v", pattern)